add basic validation of stripe keys
